+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:grouping/signup/email.dart';
 import 'package:grouping/signup/gender_n_birth.dart';
 import 'package:grouping/signup/name.dart';
 import 'package:grouping/signup/password.dart';
-import 'package:grouping/signup/validation_code.dart';
+import 'package:grouping/signup/validate_code.dart';
 import 'package:grouping/splash/main.dart';
 
 void main() => runApp(Grouping());
